requirements.txt
# AI Opportunity Index dependencies
sqlalchemy>=2.0
psycopg2-binary>=2.9
asyncpg>=0.29
alembic>=1.13
pandas>=2.0
numpy>=1.24
requests>=2.31
yfinance>=0.2.31
plotly>=5.18
sec-edgar-downloader>=5.0
pyarrow>=14.0
anthropic>=0.40
tiktoken>=0.5
beautifulsoup4>=4.12
feedparser>=6.0
lxml>=4.9
pydantic>=2.0
jinja2>=3.1
pydantic-ai>=0.1
googlesearch-python>=1.2

# Web / Subscription
litestar>=2.0
uvicorn>=0.27
stripe>=8.0
httpx>=0.27
resend>=2.0
strawberry-graphql[litestar]>=0.220
